tools/c7n_tencentcloud/pyproject.toml
[tool.poetry]
name = "c7n_tencentcloud"
version = "0.1.0"
description = "Cloud Custodian - Tencent Cloud Provider"
authors = ["Tencent Cloud"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
tencentcloud-sdk-python = "^3.0.715"
pytz = ">=2022.1"
retrying = "^1.3.3"
python-linq = "^2.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
placebo = "^0.9.0"
c7n = {path = "../..", develop = true}
freezegun = "^1.2.2"
pytest-cov = "^3.0.0"
pytest-recording = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
